data-view/requirements.txt
# requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic-settings>=2.0.0 # Handles .env loading
databases[mysql]>=0.8.0
aiomysql>=0.2.0          # MySQL driver for databases
passlib[bcrypt]>=1.7.4   # Password hashing
python-jose[cryptography]>=3.3.0 # JWT handling
kazoo>=2.9.0             # Zookeeper client
python-dotenv>=1.0.0     # Explicit dependency (though pydantic-settings can use it)
